+++ sdk/datafactory/arm-datafactory/src/models/index.ts
@@ -670,6 +670,12 @@ export interface FactoryIdentity {
   userAssignedIdentities?: { [propertyName: string]: Record<string, unknown> };
 }
 
+/** Purview configuration. */
+export interface PurviewConfiguration {
+  /** Purview resource id. */
+  purviewResourceId?: string;
+}
+
 /** Factory's git repo information. */
 export interface FactoryRepoConfiguration {
   /** Polymorphic discriminator, which specifies the different types this object can be */
@@ -2261,6 +2267,14 @@ export interface PrivateLinkConnectionState {
 export interface PrivateLinkConnectionApprovalRequest {
   /** The state of a private link connection */
   privateLinkServiceConnectionState?: PrivateLinkConnectionState;
+  /** The resource of private endpoint. */
+  privateEndpoint?: PrivateEndpoint;
+}
+
+/** Private endpoint which a connection belongs to. */
+export interface PrivateEndpoint {
+  /** The resource Id for private endpoint */
+  id?: string;
 }
 
 /** Wrapper for a collection of private link resources */
@@ -3672,6 +3686,8 @@ export type Factory = Resource & {
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
   readonly version?: string;
+  /** Purview information of the factory. */
+  purviewConfiguration?: PurviewConfiguration;
   /** Git repo information of the factory. */
   repoConfiguration?: FactoryRepoConfigurationUnion;
   /** List of parameters for factory. */
@@ -7418,6 +7434,8 @@ export type DataFlowSource = Transformation & {
 export type DataFlowSink = Transformation & {
   /** Schema linked service reference. */
   schemaLinkedService?: LinkedServiceReference;
+  /** Rejected data linked service reference. */
+  rejectedDataLinkedService?: LinkedServiceReference;
 };
 
 /** The location of azure blob dataset. */