+++ sdk/datafactory/arm-datafactory/src/models/mappers.ts
@@ -400,6 +400,21 @@ export const FactoryIdentity: coreClient.CompositeMapper = {
   }
 };
 
+export const PurviewConfiguration: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "PurviewConfiguration",
+    modelProperties: {
+      purviewResourceId: {
+        serializedName: "purviewResourceId",
+        type: {
+          name: "String"
+        }
+      }
+    }
+  }
+};
+
 export const FactoryRepoConfiguration: coreClient.CompositeMapper = {
   type: {
     name: "Composite",
@@ -3640,6 +3655,28 @@ export const PrivateLinkConnectionApprovalRequest: coreClient.CompositeMapper = 
         type: {
           name: "Composite",
           className: "PrivateLinkConnectionState"
+        }
+      },
+      privateEndpoint: {
+        serializedName: "privateEndpoint",
+        type: {
+          name: "Composite",
+          className: "PrivateEndpoint"
+        }
+      }
+    }
+  }
+};
+
+export const PrivateEndpoint: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "PrivateEndpoint",
+    modelProperties: {
+      id: {
+        serializedName: "id",
+        type: {
+          name: "String"
         }
       }
     }
@@ -7137,6 +7174,13 @@ export const Factory: coreClient.CompositeMapper = {
         readOnly: true,
         type: {
           name: "String"
+        }
+      },
+      purviewConfiguration: {
+        serializedName: "properties.purviewConfiguration",
+        type: {
+          name: "Composite",
+          className: "PurviewConfiguration"
         }
       },
       repoConfiguration: {
@@ -18468,6 +18512,13 @@ export const DataFlowSink: coreClient.CompositeMapper = {
       ...Transformation.type.modelProperties,
       schemaLinkedService: {
         serializedName: "schemaLinkedService",
+        type: {
+          name: "Composite",
+          className: "LinkedServiceReference"
+        }
+      },
+      rejectedDataLinkedService: {
+        serializedName: "rejectedDataLinkedService",
         type: {
           name: "Composite",
           className: "LinkedServiceReference"